auth/app/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.47.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
gunicorn==20.1.0
djangorestframework-api-key==2.*